html/semantics/forms/the-input-element/week.html

<!DOCTYPE html>
<html>
  <head>
    <title>Week input element</title>
    <link rel="author" title="Shaoning Zeng" href="mailto:shaoning.zeng@gmail.com">
    <link rel="help" href="http://www.w3.org/TR/html5/forms.html#week-state-(type=week)">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <h1>Week input element</h1>
    <div style="display: none">

    <input id="week" type="week" value="2013-W45" />

    <input id="week_bad0" type="week" />
    <input id="week_bad1" type="week" value="45" />
    <input id="week_bad2" type="week" value="2013" />
    <input id="week_bad3" type="week" value="2013-W99" />

    <input id="week_min" type="week" min="2013-W40" />
    <input id="week_min_bad1" type="week" min="20" />
    <input id="week_min_bad2" type="week" min="2013" />
    <input id="week_min_bad3" type="week" min="2013-W400" />

    <input id="week_max" type="week" max="2013-W40" />
    <input id="week_max_bad1" type="week" max="20" />
    <input id="week_max_bad2" type="week" max="2013" />
    <input id="week_max_bad3" type="week" max="2013-W400" />

    </div>
    <div id="log"></div>
  <script type="text/javascript">

    var element = document.getElementById("week");
    test(
      function() {
        assert_equals(element.type, "week");
        assert_equals(element.value, "2013-W45");
      }, "The value attribute, if specified and not empty, must have a value that is a valid week string.");

    test(
      function() {
        assert_equals(document.getElementById("week_bad0").value, '');
        assert_equals(document.getElementById("week_bad1").value, '');
        assert_equals(document.getElementById("week_bad2").value, '');
        assert_equals(document.getElementById("week_bad3").value, '');
      }, "The value sanitization algorithm is as follows: If the value of the element is not a valid week string, then set it to the empty string instead.");

    test(
      function() {
        assert_equals(document.getElementById("week_min").min, '2013-W40');
        assert_equals(document.getElementById("week_min_bad1").min, "");
        assert_equals(document.getElementById("week_min_bad2").min, "");
        assert_equals(document.getElementById("week_min_bad3").min, "");
      }, "The min attribute, if specified, must have a value that is a valid week string.");

    test(
      function() {
        assert_equals(document.getElementById("week_max").max, '2013-W40');
        assert_equals(document.getElementById("week_max_bad1").max, "");
        assert_equals(document.getElementById("week_max_bad2").max, "");
        assert_equals(document.getElementById("week_max_bad3").max, "");
      }, "The max attribute, if specified, must have a value that is a valid week string.");

  </script>
  </body>
</html>
